Add admin endpoint to bypass npub verification
Some users might request to pay for a subscription just to get a low
subscription number, even if they do not have a compatible Damus
version.

Usually it is impossible for them to complete the checkout process
without a compatible Damus version.

However, some users might not be able to get the latest version but
still want a Purple membership for the badge.

This commit adds a special password protected endpoint (Password set in ENV) that allows a checkout object to be created with the verification step bypassed.

Testing 1: Ran LN flow locally successfully
Testing 2: Sent CURL command to create a checkout object for a specific pubkey. Then loaded the checkout id into the checkout page, and paid the invoice.

Signed-off-by: Daniel D’Aquino <daniel@daquino.me>
Link: 20240131030626.94872-2-daniel@daquino.me
Signed-off-by: William Casarin <jb55@jb55.com>