Add type check tooling
This adds typescript tooling for type checking during development, as an
extra tool to help debugging and improve robustness of Javascript code

This does not add a compilation step or any dependencies or changes on
existing node.js environment.

It only adds an npm script that developers can run locally to do a type
check.

Type checks can be done on JS files as they are, or improved with the
use of `// @ts-check` and JSDoc comments with types documented

Signed-off-by: Daniel D’Aquino <daniel@daquino.me>
Reviewed-by: William Casarin <jb55@jb55.com>
Signed-off-by: William Casarin <jb55@jb55.com>