Add test/admin endpoint for overriding the expiry date
This commit adds a test admin endpoint for overriding the expiry date of
a user. This was created to facilitate testing of impending purple
expiration notifications, but could come in handy in other situations.

Testing: Sent a request to this endpoint and successfully changed the expiry date of a test user.
Signed-off-by: Daniel D’Aquino <daniel@daquino.me>