Fix crash when check-invoice is called when there is no invoice
During test and development of website improvements I noticed that
calling the endpoint to check-invoice when there is no invoice or
checkout object would cause a crash.

This commit fixes this

Testing: Performing hot-reloading of checkout page in the same
preconditions that triggered the bug. Also, reran automated tests to
ensure there are no regressions

Signed-off-by: Daniel D’Aquino <daniel@daquino.me>