Add Order ID lookup functions to aid customer support
This commit adds lookup functions that allows transaction histories to
be found using the Order ID from the customer. The Order ID is a string
that gets included in each email receipt and can be used in customer
support situations.

Currently the function is not exposed in any external APIs. It can only
be used interactively on a Node.js console.

From the perspective of the API and the server functionality, this commit should be a no-op.

-------
Testing
-------

Damus API: This commit
Coverage:
- All automated tests are passing (with no `.env`)
- Typescript type check has no regressions
- New function was tested by using it via Node.js console to help solve a real customer support request

Closes: https://github.com/damus-io/api/issues/9
Changelog-Added: Add Order ID lookup functions to aid customer support
Signed-off-by: Daniel D’Aquino <daniel@daquino.me>
Reviewed-by: William Casarin <jb55@jb55.com>