Fix external test dependency
It was found that the automated tests were calling an external Apple
API, and those calls were seen to fail. Since this brings a test
reliability issue, that portion of the Apple IAP library was mocked to
remove external HTTP calls and improve robustness of the test.

Furthermore, some mock data, test controller APIs, and a test had to be
adapted to work with the new changes.

Changelog-Fixed: Fix external test dependency on an Apple API call.
Signed-off-by: Daniel D’Aquino <daniel@daquino.me>
Reviewed-by: William Casarin <jb55@jb55.com>